finishing splitting up gui
